Many big changes (lazy commit due to period of time), debug build can now run again
Months of previous changes I've long forgotten, and a few updates to compiler definitions such as _DEBUG (which is used by imgui) to PIXEL_PROJECT_DEBUG which strips out debug messages.

Removal of fmt which was causing problems, using modern c++ std::format which does most of the same, plus the tweaks to make this change work with project.

+++ PixelProject/Scripts/Math/Vec2/SVec2.cpp
@@ -12,7 +12,7 @@ SVec2::SVec2(const short x, const short y): x(x), y(y)
 
 SVec2 SVec2::operator+(const SVec2& rhs) const
 {
-   return {x + rhs.x, y + rhs.y};
+   return {static_cast<short>(x + rhs.x), static_cast<short>(y + rhs.y)};
 }
 
 SVec2& SVec2::operator=(const SVec2 other)
